Time: 2 ms (52.81%), Space: 7.7 MB (29.35%) - LeetHub

+++ 0790-domino-and-tromino-tiling/0790-domino-and-tromino-tiling.cpp
@@ -33,6 +33,6 @@ public:
             part[i] = (part[i-1] + full[i - 2]) % MOD; 
         }
 
-        return static_cast<int>(full[n]);
+        return full[n]; 
     }
 };
